+++ src/DICOMImage.cpp
@@ -229,12 +229,12 @@ ImagePlane* DICOMImage::GetImagePlaneFromDICOMHeaderInfo() const
 	// rather than centre; (0020, 0032) is the centre of the first voxel
 	plane_->tlc = position3D_ - 0.5 * pixelSizeX_ * orientation1_ -  0.5f * pixelSizeY_ * orientation2_;
 
-	float fieldOfViewX = width_ * pixelSizeX_;//JDCHUNG consider name change
+	double fieldOfViewX = width_ * pixelSizeX_;//JDCHUNG consider name change
 	cout << "width in mm = " << fieldOfViewX ;
 	
 	plane_->trc = plane_->tlc + fieldOfViewX * orientation1_;
 
-	float fieldOfViewY = height_ * pixelSizeY_;//JDCHUNG
+	double fieldOfViewY = height_ * pixelSizeY_;//JDCHUNG
 	cout << ", height in mm = " << fieldOfViewY << endl ;
 	
 	plane_->blc = plane_->tlc + fieldOfViewY * orientation2_;
